// apps/web/app/(main)/labs/schemas.ts
import * as z from "zod";
import { EnumLabStatus } from "./enum";

export const labFormSchema = z.object({
  title: z.string().optional(), // Title will be generated by AI
  description: z.string().optional(),
  estimatedTime: z.string().optional(),
  difficulty: z.enum(["intro", "intermediate", "advanced"]).optional(), // Will be generated by AI
  labType: z.enum(["concept", "practice", "exploration"]).optional(),
  status: z.enum(Object.values(EnumLabStatus) as [EnumLabStatus, ...EnumLabStatus[]]).optional() // Will be set by backend
});

export type LabFormValues = z.infer<typeof labFormSchema>;